+++ packages/core/src/tools/read.ts
@@ -1,5 +1,5 @@
 export { getComponents } from './read/components'
-export { listAvailableFonts, listFonts } from './read/fonts'
+export { getFontStatus, listAvailableFonts, listFonts } from './read/fonts'
 export { diffJSX, getJSX } from './read/jsx'
 export { findNodes, getNode, getPageTree } from './read/nodes'
 export { getCurrentPage, listPages, pageBounds, switchPage } from './read/pages'